+++ projects/model/src/extensions/builtin_types/model/casters/hash.rb
@@ -14,7 +14,13 @@ module Foobara
           end
 
           def model_class
-            Object.const_get(parent_declaration_data[:model_class])
+            type = parent_declaration_data[:type]
+
+            if type == :model
+              Object.const_get(parent_declaration_data[:model_class])
+            else
+              Foobara::Namespace.current.foobara_lookup_type!(type).target_class
+            end
           end
         end
       end