+++ spec/foobara/command/entity_inputs_spec.rb
@@ -46,8 +46,9 @@ RSpec.describe "Entity inputs for commands" do
 
     stub_class :CreateFan, Foobara::Command do
       inputs_type_declaration = Foobara::Util.deep_dup(Fan.attributes_type.declaration_data)
-      inputs_type_declaration[:element_type_declarations][:fan_of][:element_type_declaration] =
-        { type: :User, mutable: ["fan_count"] }
+      element_type_declarations = inputs_type_declaration[:element_type_declarations]
+      element_type_declarations[:fan_of][:element_type_declaration] = { type: :User, mutable: ["fan_count"] }
+      element_type_declarations[:owner][:mutable] = false
 
       inputs inputs_type_declaration
       result Fan